refactor(genie): split the deterministic guardrail battery out of the router
backend/api/genie.py crossed the 1000-line monolith gate (891 on main -> 1349)
when the async lifecycle landed, and the allowlist covering it expired on
2026-07-31, so the gate had been failing for every file still listed.

Move the whole pre-Genie decision path -- protected-class, instruction-override,
outreach, PII, scope-bypass, source-gap, off-topic, cross-lender, sales-ops and
footprint guards, refusal shaping, and the governed output block -- verbatim
into backend/services/genie_deterministic.py. It is policy rather than routing,
so it belongs beside the other Genie services, and the router drops to 734
lines: its allowlist entry is removed rather than re-dated.

The move is byte-identical and the router re-imports the names it already
called, so route bodies are untouched. That matters twice: the audit
source-contract test inspects those bodies for _required_audit_write /
_deterministic_genie_response(, and the OpenAPI baseline pins the route
docstrings. Three test files that reached into the router for guardrail
aliases or monkeypatched sales_ops_genie_response now point at the new module.

For the eight files still oversize, record the dated schedule decision the plan
document requires instead of a silent extension: new expiry 2026-09-15, plus
split boundaries for backend/api/outreach.py, which this branch took from 938
to 2192 lines and which is left intact here because it is the human-approval
and audit path. A file may be re-dated at most once more.
